back/accounts/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.6
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
pycparser==2.22
pygame==2.5.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1

